Implementacion de los menus superiores para logout
