+++ app/src/main/java/com/example/whereyou/MapaGrupoActivity.kt
@@ -4,9 +4,11 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.whereyou.databinding.ActivityMapaGrupoBinding
+import com.google.firebase.auth.FirebaseAuth
 
 class MapaGrupoActivity : AppCompatActivity() {
     private lateinit var binding : ActivityMapaGrupoBinding
+    private lateinit var auth: FirebaseAuth
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMapaGrupoBinding.inflate(layoutInflater)
@@ -18,7 +20,7 @@ class MapaGrupoActivity : AppCompatActivity() {
             }
         }
 
-        binding.chats.setOnClickListener{
+        binding.HAChats.setOnClickListener{
             startActivity(Intent(baseContext, ChatsActivity::class.java))
         }
 
@@ -43,6 +45,12 @@ class MapaGrupoActivity : AppCompatActivity() {
 
         binding.HAPerfil.setOnClickListener {
             startActivity(Intent(baseContext,PerfilActivity::class.java))
+        }
+        binding.GAMenuLogOut.setOnClickListener {
+            auth.signOut()
+            val i = Intent(this, LoginActivity::class.java)
+            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            startActivity(i)
         }
     }
 }
