+++ app/src/main/res/layout/activity_crear_grupo.xml
@@ -7,55 +7,45 @@
     tools:context=".CrearGrupoActivity">
 
     <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbarCrearGrupo"
+        android:id="@+id/toolbarNotificaciones"
         android:layout_width="match_parent"
         android:layout_height="60dp"
         android:background="@color/azul3D"
         android:minHeight="?attr/actionBarSize"
         android:theme="?attr/actionBarTheme"
-        app:layout_constraintBottom_toTopOf="@+id/constraintL"
+        app:layout_constraintBottom_toTopOf="@+id/linearL"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageButton
-        android:id="@+id/GAOpciones"
+        android:id="@+id/GAMenuLogOut"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:background="@android:color/transparent"
         android:padding="8dp"
-        android:src="@drawable/menu_toolbar"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarCrearGrupo"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarCrearGrupo" />
-
-    <ImageButton
-        android:id="@+id/GAInformacion"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:padding="8dp"
-        android:src="@drawable/info_toolbar"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarCrearGrupo"
+        android:src="@drawable/baseline_logout_24"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarCrearGrupo" />
+        app:layout_constraintStart_toEndOf="@+id/Perfil"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
     <TextView
         android:id="@+id/Perfil"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
-        android:gravity="center"
-        android:text="Crear Grupo"
+        android:layout_marginStart="16dp"
+        android:gravity="left"
+        android:text="@string/crear_grupo"
         android:textColor="#000000"
         android:textSize="18sp"
         android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarCrearGrupo"
-        app:layout_constraintEnd_toStartOf="@id/GAInformacion"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
+        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@id/GAOpciones"
-        app:layout_constraintTop_toTopOf="@id/toolbarCrearGrupo" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
 
     <LinearLayout
@@ -181,16 +171,7 @@
 
 
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraintL"
-        android:layout_width="0dp"
-        android:layout_height="70dp"
-        android:background="@color/gray"
-        app:layout_constraintBottom_toTopOf="@+id/linearL"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbarCrearGrupo"
-        app:layout_constraintVertical_bias="0.0"/>
+
 
     <LinearLayout
         android:id="@+id/linearL"
@@ -203,7 +184,7 @@
         app:layout_constraintBottom_toTopOf="@+id/separador"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/constraintL">
+        app:layout_constraintTop_toBottomOf="@id/toolbarNotificaciones">
 
         <ImageView
             android:id="@+id/imageView"
