+++ app/src/main/res/layout/activity_group_chat.xml
@@ -19,43 +19,34 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageButton
-        android:id="@+id/GAOpciones"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:background="@android:color/transparent"
-        android:padding="8dp"
-        android:src="@drawable/menu_toolbar"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
-
-    <ImageButton
-        android:id="@+id/GAInformacion"
+        android:id="@+id/GAMenuLogOut"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:background="@android:color/transparent"
         android:padding="8dp"
-        android:src="@drawable/info_toolbar"
+        android:src="@drawable/baseline_logout_24"
         app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/Perfil"
         app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
     <TextView
-        android:id="@+id/titleText"
+        android:id="@+id/Perfil"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
         android:layout_marginEnd="16dp"
-        android:gravity="center"
-        android:text="Grupo Chat"
+        android:layout_marginStart="16dp"
+        android:gravity="left"
+        android:text="@string/chat_del_grupo"
         android:textColor="#000000"
         android:textSize="18sp"
         android:textStyle="bold"
         app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toStartOf="@id/GAInformacion"
+        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@id/GAOpciones"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
+
 
 
     <LinearLayout
