+++ app/src/main/res/layout/activity_grupos.xml
@@ -7,47 +7,47 @@
     tools:context=".GruposActivity">
 
 
-    <LinearLayout
-        android:id="@+id/topMenu"
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbarNotificaciones"
         android:layout_width="match_parent"
         android:layout_height="60dp"
         android:background="@color/azul3D"
-        android:orientation="horizontal"
-        android:padding="5dp"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="?attr/actionBarTheme"
+        app:layout_constraintBottom_toTopOf="@+id/GA_buscador"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toTopOf="parent" />
 
-        <ImageButton
-            android:id="@+id/GAOpciones"
-            android:layout_width="20dp"
-            android:layout_height="50dp"
-            android:layout_weight="1"
-            android:background="@color/azul3D"
-            android:src="@drawable/baseline_view_headline_24" />
+    <ImageButton
+        android:id="@+id/GAMenuLogOut"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@android:color/transparent"
+        android:padding="8dp"
+        android:src="@drawable/baseline_logout_24"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/Perfil"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_weight="2"
-            android:gravity="center"
-            android:text="Nombre Grupo"
-            android:textColor="@color/black"
-            android:textSize="18sp"
-            android:textStyle="bold" />
-
-        <ImageButton
-            android:id="@+id/GAInformacion"
-            android:layout_width="20dp"
-            android:layout_height="50dp"
-            android:layout_weight="1"
-            android:backgroundTint="@color/azul3D"
-            android:src="@drawable/info_toolbar" />
-        />
-
-    </LinearLayout>
+    <TextView
+        android:id="@+id/Perfil"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="16dp"
+        android:layout_marginStart="16dp"
+        android:gravity="left"
+        android:text="@string/grupos"
+        android:textColor="#000000"
+        android:textSize="18sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
+        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
 
 
@@ -56,13 +56,12 @@
         android:layout_width="0dp"
         android:layout_height="50dp"
         android:layout_marginBottom="6dp"
-        android:ems="10"
         android:inputType="text"
         android:text="Nombre de grupo o persona"
         app:layout_constraintBottom_toTopOf="@+id/GA_listaGrupos"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/topMenu" />
+        app:layout_constraintTop_toBottomOf="@id/toolbarNotificaciones" />
 
     <ListView
         android:id="@+id/GA_listaGrupos"
@@ -70,6 +69,7 @@
         android:layout_height="454dp"
         app:layout_constraintBottom_toTopOf="@+id/GA_nuevoContacto"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/GA_buscador" />
