+++ app/src/main/res/layout/activity_mapa_grupo.xml
@@ -6,66 +6,47 @@
     android:layout_height="match_parent"
     tools:context=".MapaGrupoActivity">
 
-    <LinearLayout
-        android:id="@+id/topMenu"
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbarNotificaciones"
         android:layout_width="match_parent"
         android:layout_height="60dp"
-        android:padding="5dp"
-        android:orientation="horizontal"
+        android:background="@color/azul3D"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="?attr/actionBarTheme"
+        app:layout_constraintBottom_toTopOf="@+id/mapSeccion"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        android:background="@color/azul3D"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageButton
-            android:id="@+id/GAOpciones"
-            android:layout_width="20dp"
-            android:layout_height="50dp"
-            android:layout_weight="1"
-            android:background="@color/azul3D"
-            android:src="@drawable/baseline_view_headline_24" />
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="2"
-            android:layout_gravity="center"
-            android:textColor="@color/black"
-            android:textStyle="bold"
-            android:textSize="18sp"
-            android:text="Nombre Grupo" />
+        app:layout_constraintTop_toTopOf="parent" />
 
 
-        <ImageButton
-            android:id="@+id/chats"
-            android:layout_width="28dp"
-            android:layout_height="46dp"
-            android:layout_gravity="center"
-            android:layout_weight="1"
-            android:backgroundTint="@color/azul3D"
-            android:scaleType="fitCenter"
-            android:src="@drawable/chats" />
+    <ImageButton
+        android:id="@+id/GAMenuLogOut"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@android:color/transparent"
+        android:padding="8dp"
+        android:src="@drawable/baseline_logout_24"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/Perfil"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
-
-        <ImageButton
-            android:id="@+id/GAInformacion"
-            android:layout_width="20dp"
-            android:layout_height="50dp"
-            android:layout_weight="1"
-            android:backgroundTint="@color/azul3D"
-            android:src="@drawable/baseline_notifications_24" />
-
-        <Switch
-            android:id="@+id/switchUbicacion"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="5"
-            android:textColor="#343f4b"
-            android:layout_gravity="center"
-            android:text="Ubicación" />
-
-    </LinearLayout>
+    <TextView
+        android:id="@+id/Perfil"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="16dp"
+        android:layout_marginStart="16dp"
+        android:gravity="left"
+        android:text="Grupo: Grupo Prueba"
+        android:textColor="#000000"
+        android:textSize="18sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
+        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
 
     <LinearLayout
@@ -293,12 +274,13 @@
     </LinearLayout>
 
     <FrameLayout
+        android:id="@+id/mapSeccion"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/peopleMenu"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/topMenu">
+        app:layout_constraintTop_toBottomOf="@+id/toolbarNotificaciones">
