+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,9 @@
     <string name="app_name">WhereYou</string>
     <string name="hintContrasenia">Contraseña</string>
     <string name="hintUsuario">Usuario o Email</string>
+    <string name="perfil">Perfil</string>
+    <string name="notificaciones">Notificaciones</string>
+    <string name="grupos">Grupos</string>
+    <string name="chat_del_grupo">Chat del grupo</string>
+    <string name="crear_grupo">Crear Grupo</string>
 </resources>
