+++ app/src/main/res/values/themes.xml
@@ -1,8 +1,8 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.WhereYou" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.WhereYou" parent="@style/Theme.AppCompat.Light.DarkActionBar">
         <!-- Customize your light theme here. -->
-        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <item name="colorPrimary">@color/azul68</item>
     </style>
 
     <style name="Theme.WhereYou" parent="Base.Theme.WhereYou" />
